packages/services/requirements.txt
aniso8601==9.0.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
decorator==5.0.7
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
itsdangerous==2.0.0
jedi==0.18.0
Jinja2==3.0.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lxml==4.6.3
MarkupSafe==2.0.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.0
